Use `-ff` and `-watermodel` options of pdb2gmx
By default, the forcefield and water models are selected
interactively, which is not ideal in scripts. Previously,
the problem was solved by feeding the selected options in
the script. However, this is not robust since the option
number can change if one e.g. adds another forcefield to
the GROMACS installation. This makes use of the `-ff` and
`-watermodel` options of pdb2gmx to select the forcefield.

Note that the `-watermodel` option takes enumerated value,
which is then used to look for a specific file in the
forcefield folder. In the trappeua, the tip4p water model
is called `tip4p2005`, not `tip4p`. Which will have to be
changed so that pdb2gmx will be able to find the water
model itp file. See the change to the documentation here.

Resolves issue #1